+++ c/tskit/trees.h
@@ -70,6 +70,8 @@ typedef struct {
     bool time_uncalibrated;
     /* Are all genome coordinates discrete? */
     bool discrete_genome;
+    /* Are all time values discrete? */
+    bool discrete_time;
     /* Breakpoints along the sequence, including 0 and L. */
     double *breakpoints;
     /* If a node is a sample, map to its index in the samples list */
@@ -266,6 +268,7 @@ const tsk_id_t *tsk_treeseq_get_samples(const tsk_treeseq_t *self);
 const tsk_id_t *tsk_treeseq_get_sample_index_map(const tsk_treeseq_t *self);
 bool tsk_treeseq_is_sample(const tsk_treeseq_t *self, tsk_id_t u);
 bool tsk_treeseq_get_discrete_genome(const tsk_treeseq_t *self);
+bool tsk_treeseq_get_discrete_time(const tsk_treeseq_t *self);
 
 int tsk_treeseq_get_node(const tsk_treeseq_t *self, tsk_id_t index, tsk_node_t *node);
 int tsk_treeseq_get_edge(const tsk_treeseq_t *self, tsk_id_t index, tsk_edge_t *edge);