+++ python/CHANGELOG.rst
@@ -86,6 +86,9 @@
 
 - Add the ``TreeSequence.alignments`` method. (user:`jeromekelleher`, :pr:`1825`)
 
+- Add the ``discrete_time`` property to the TreeSequence class which is true if
+  all time coordinates are discrete or unknown (:user:`benjeffery`, :issue:`1839`, :pr:`1890`)
+
 **Fixes**
 
 - `dump_tables` omitted individual parents. (:user:`benjeffery`, :issue:`1828`, :pr:`1884`)
@@ -100,6 +103,7 @@
   sample labelling pattern as ``as_newick``.
   (:user:`jeetsukumaran`, :user:`jeromekelleher`, :issue:`1785`, :pr:`1835`,
   :pr:`1836`, :pr:`1838`)
+
 
 --------------------
 [0.3.7] - 2021-07-08